Perform renames in TASRSP
