+++ src/Plugins.Win32.TASRSP/Config.hpp
