+++ src/Plugins.Win32.TASRSP/HLE.hpp
