+++ src/Plugins.Win32.TASRSP/MP3.cpp
@@ -4,8 +4,8 @@
  * SPDX-License-Identifier: GPL-2.0-or-later
  */
 
-#include "Main.h"
-#include "HLE.h"
+#include "Main.hpp"
+#include "HLE.hpp"
 
 static uint16_t DeWindowLUT[0x420] = {
     0x0000, 0xFFF3, 0x005D, 0xFF38, 0x037A, 0xF736, 0x0B37, 0xC00E, 0x7FFF, 0x3FF2, 0x0B37, 0x08CA, 0x037A, 0x00C8,
