+++ src/Plugins.Win32.TASRSP/Main.hpp
