--- src/Plugins.Win32.TASRSP/stdafx.cpp
@@ -1,7 +1,0 @@
-/*
- * Copyright (c) 2026, hacktarux-azimer-rsp-hle maintainers, contributors, and original authors (Hacktarux, Azimer).
- *
- * SPDX-License-Identifier: GPL-2.0-or-later
- */
-
-#include "stdafx.h"
