--- src/Plugins.Win32.TASRSP/stdafx.h
@@ -1,17 +1,0 @@
-/*
- * Copyright (c) 2026, hacktarux-azimer-rsp-hle maintainers, contributors, and original authors (Hacktarux, Azimer).
- *
- * SPDX-License-Identifier: GPL-2.0-or-later
- */
-
-#pragma once
-
-#include "resource.h"
-#include <Windows.h>
-#include <cassert>
-#include <cstdio>
-#include <filesystem>
-#include <shlobj.h>
-#include <string>
-#include <vector>
-#include "core_plugin.h"
